Reuse i18nResultsCount function in header and content link
